feat: refactor importer command and create a base for adding sbom importer
